server/requirements.txt
# Chotto Voice Server - Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.26.0
libsql-experimental>=0.0.47
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
openai>=1.12.0
stripe>=7.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
